requirements.txt
requests
pyyaml
lxml
pycurl
bibtexparser